DiskCache: Make the cache age a long value
This is to support larger ages. Adjust the code in `MavenSupport` to use
a more readable duration extension function to specify the age.

+++ utils/common/src/main/kotlin/DiskCache.kt
@@ -45,7 +45,7 @@ class DiskCache(
     /**
      * Duration in seconds that cache entries are valid.
      */
-    private val maxCacheEntryAgeInSeconds: Int
+    private val maxCacheEntryAgeInSeconds: Long
 ) {
     companion object : Logging {
         const val INDEX_FULL_KEY = 0
